+++ src/service/user-service-interface.ts
@@ -1,4 +1,4 @@
-import User from '../models/user';
+import User from '../model/user';
 
 export default interface UserServiceInterface {
     getUser(username: string): Promise<User>;
